+++ src/linkedlist/list.h
@@ -9,13 +9,13 @@ struct Node {
 typedef struct LinkedList {
   struct Node *head;
   struct Node *tail;
-  int size;
+  long long size;
 } LinkedList;
 
 struct LinkedList *createlist();
 void *deletelist(struct LinkedList *list);
 void *addtolist(struct LinkedList *list, void *value);
-void *getfromindex(struct LinkedList *list, int index);
-void *removefromindex(struct LinkedList *list, int index);
+void *getfromindex(struct LinkedList *list, long long index);
+void *removefromindex(struct LinkedList *list, long long index);
 
 #endif // ARIA_DATA_STRUCTURES_LIST_H
